requirements.txt
# PyTorch and Computer Vision
torch==2.1.2
torchvision==0.16.2

# Data Processing
pandas>=2.0.0

# Image Processing
Pillow>=10.0.0

# Utilities
numpy>=1.24.0  # Required by torch/torchvision
